refactor(CouponModal): 선택된 쿠폰 상태 커스텀 훅으로 분리
