+++ src/hooks/useCouponCalculator.ts
@@ -5,19 +5,19 @@ import { sortCouponsByDiscountRate } from '../utils/couponSorter';
 import { calculateDiscountPrice } from '../utils/couponCalculator';
 import { Coupon } from '../types/Coupon.type';
 
-const useCartCalculator = () => {
+const useCouponCalculator = () => {
   const selectedCartItemList = useRecoilValue(selectedCartItemListState);
   const totalOrderPrice = useRecoilValue(totalOrderPriceSelector);
   const deliveryFee = useRecoilValue(deliveryFeeSelector);
 
-  const calculateTotalDiscoutPrice = (coupons: Coupon[]) => {
+  const calculateTotalDiscountPrice = (coupons: Coupon[]) => {
     return sortCouponsByDiscountRate(coupons).reduce(
       (acc, cur) => (acc += calculateDiscountPrice(cur, selectedCartItemList, deliveryFee, totalOrderPrice)),
       0,
     );
   };
 
-  return { calculateTotalDiscoutPrice };
+  return { calculateTotalDiscountPrice };
 };
 
-export default useCartCalculator;
+export default useCouponCalculator;
